+++ R/step.metrics.R
@@ -23,8 +23,10 @@
 #'   `meta/ms2.out/`.
 #' @param outputdir Character. Directory where results should be stored.
 #'   Subfolders will be created as needed (`daySummary/`).
-#' @param idloc Character (default = `"_"`). Delimiter used to extract
+#' @param idloc Character (default = NULL). Delimiter used to extract
 #'   participant IDs from filenames (ID is expected before this string).
+#'   If working with GGIR output, `idloc` is ignored and the GGIR-identified
+#'   ID is directly used to ensure matching. 
 #' @param cadence_bands Numeric vector (default =
 #'   `c(0, 1, 20, 40, 60, 80, 100, 120, Inf)`).
 #'   Breakpoints (in steps/min) used to compute time and steps per cadence band.
@@ -81,7 +83,7 @@
 #' @importFrom utils write.csv
 #' @export
 step.metrics = function(datadir, outputdir="./",
-                        idloc = "_",
+                        idloc = NULL,
                         cadence_bands = c(0, 1, 20, 40, 60, 80, 100, 120, Inf),
                         cadence_peaks = c(1, 30, 60),
                         cadence_MOD = 100,
@@ -104,10 +106,20 @@ step.metrics = function(datadir, outputdir="./",
 
   # Get IDs -----
   ids = c()
-  for (i in 1:length(files)) {
-    ids[i] = unlist(strsplit(files[i], split = idloc, fixed = TRUE))[1]
+  # identify ids as a way to detect if any recording is splitted in several files
+  if (is.null(idloc) || isGGIR) {
+    # Use the full file name as the ID if no idloc specified or if GGIR 
+    #   (ID will be derived from GGIR milestone file)
+    ids = files
+  } else {
+    for (i in 1:length(files)) {
+      # Use the provided idloc delimiter
+      ids[i] = unlist(strsplit(files[i], split = idloc, fixed = TRUE))[1]
+    }
   }
   ids = unique(ids)
+
+  # This handles the removal of the .RData extension from GGIR files
   ids = gsub(".RData$", "", ids)
 
   if (verbose == TRUE) {
@@ -117,12 +129,13 @@ step.metrics = function(datadir, outputdir="./",
   }
 
   #Loop through the files
-  if (verbose == TRUE) cat("Processing participant: ")
   for (i in 1:length(ids)) {
-    if (verbose == TRUE) cat(paste0(ids[i], " "))
+    if (verbose == TRUE) cat(paste0("\rRecording ", i, "/", length(ids), ": ", ids[i]))
     # read data ----
-    files2read = grep(ids[i], files_fn, value = TRUE)
+    files2read = grep(ids[i], files_fn, value = TRUE, fixed = TRUE)
     data = readFile(files2read, time_format = time_format)
+    id_ggir = data$id
+    data = data$data
 
     # create vector with day indices -----
     day = define_day_indices(data$timestamp)
@@ -131,7 +144,11 @@ step.metrics = function(datadir, outputdir="./",
     for (di in 1:length(unique(day))) {
 
       # ID
-      id = ids[i]
+      if (isGGIR) {
+        id = id_ggir
+      } else {
+        id = ids[i]
+      }
 
       #Date
       if (di == 1) date = wday = wday_num = c()
@@ -262,9 +279,9 @@ step.metrics = function(datadir, outputdir="./",
   colnames(output) = names.out.2
 
   #Loop through files to calculate mean variables
-  if (verbose == TRUE) cat("Processing participant: ")
   for (i in 1:length(files)) {
-    if (verbose == TRUE) cat(gsub("_DaySum.csv", "", files[i]), " ")
+    if (verbose == TRUE) cat(paste0("\rID ", i, "/", length(files), ": ", 
+                                    gsub("_DaySum.csv", "", files[i])))
     D = read.csv(paste0(outputdir,"/daySummary/", files[i]))
     if (isGGIR == TRUE) {
       if (is.null(includeawakecrit)) {